+++ internal/store/sqlc/models.go
@@ -69,6 +69,7 @@ type Collector struct {
 	Role      string             `json:"role"`
 	CreatedAt pgtype.Timestamptz `json:"created_at"`
 	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
+	Labels    json.RawMessage    `json:"labels"`
 }
 
 type CollectorInstance struct {